reset snapshotted props when needed

+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/BinlogReader.java
@@ -7,7 +7,9 @@ package io.debezium.connector.mysql;
 
 import java.io.IOException;
 import java.io.Serializable;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.EnumMap;
 import java.util.HashMap;
@@ -44,7 +46,9 @@ import com.github.shyiko.mysql.binlog.network.SSLMode;
 
 import io.debezium.connector.mysql.MySqlConnectorConfig.SecureConnectionMode;
 import io.debezium.connector.mysql.RecordMakers.RecordsForTable;
+import io.debezium.document.Array;
 import io.debezium.function.BlockingConsumer;
+import io.debezium.jdbc.JdbcConnection;
 import io.debezium.relational.TableId;
 import io.debezium.util.Clock;
 import io.debezium.util.ElapsedTimeStrategy;
@@ -230,7 +234,7 @@ public class BinlogReader extends AbstractReader {
         }
     }
 
-    private AtomicBoolean threadKilled = new AtomicBoolean(false);
+    private final AtomicBoolean threadKilled = new AtomicBoolean(false);
     @Override
     protected void doStop() {
         try {
@@ -252,9 +256,19 @@ public class BinlogReader extends AbstractReader {
     protected void doCleanup() {
         try {
             if (!threadKilled.get()) {
-                logger.info("Kill binlog dump thread {}. ", client.getConnectionId());
-                context.jdbc().execute("kill " + client.getConnectionId());
-                threadKilled.compareAndSet(false, true);
+                List<Long> ids = new ArrayList<>();
+                context.jdbc.query("show processlist;", rs -> {
+                    while (rs.next()) {
+                        ids.add(Long.parseLong(rs.getString("Id")));
+                    }
+                });
+                if (ids.contains(client.getConnectionId())) {
+                    context.jdbc().execute("kill " + client.getConnectionId());
+                    logger.info("Binlog dump thread {} killed. ", client.getConnectionId());
+                    threadKilled.compareAndSet(false, true);
+                } else {
+                    logger.info("Binlog dump thread {} does not exist.", client.getConnectionId());
+                }
             }
         } catch (SQLException e) {
             logger.error("kill binlog dump thread {} failed. ", client.getConnectionId());
